Remove unused functions + rename files
